Server file for this application
This server file is written in NodeJS and the users need to install `nodejs`, `npm`, `npm socket.io express`, and then start the server using the command: ```node chat_server.js```. And once a server connected, start and android app and join the room and start chatting. This server will not be working with a physical device because server is running localhost with client-side IP: 10.0.2.2:5000. Here post number is important developers should maintain the endpoint and the port number properly. That's it